requirements.txt
 
numpy==1.18.2
tqdm==4.43.0
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.8.0+cpu
torchvision==0.9.0+cpu
opencv_contrib_python==4.2.0.34
scipy==1.4.1
scikit_image==0.15.0
Pillow==7.1.2
